set up delivery of unread message per user

+++ backend/db/crud.py
@@ -68,8 +68,8 @@ def get_messages_from_chat(db: Session, to_chat: int, skip: int, limit: int):
     return db.query(models.Message).filter(models.Message.toId == to_chat.id).order_by(models.Message.timestamp.desc()).offset(skip).limit(limit).all()
 
 
-def create_unread_message(db: Session, message_id: int):
-    db_unread_message = models.UnreadMessage(message_id=message_id)
+def create_unread_message(db: Session, message_id: int, user_id: int):
+    db_unread_message = models.UnreadMessage(message_id=message_id, user_id=user_id)
     db.add(db_unread_message)
     db.commit()
     db.refresh(db_unread_message)
@@ -80,8 +80,8 @@ def get_unread_messages(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.Message).join(models.UnreadMessage).offset(skip).limit(limit).all()
 
 
-def delete_unread_message(db: Session, message_id: int):
-    db.query(models.UnreadMessage).filter(models.UnreadMessage.message_id == message_id).delete()
+def delete_unread_message(db: Session, message_id: int, user_id: int):
+    db.query(models.UnreadMessage).filter(models.UnreadMessage.message_id == message_id).filter(models.UnreadMessage.user_id == user_id).delete()
     db.commit()
     return message_id
